LayoutTests/fast/frames/seamless/seamless-nested.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<div id="parent" style="width: 200px; height: 200px;">
<iframe id="iframe" seamless src="resources/nested-seamless.html"></iframe>
</div>
<script>
debug("Test that nested seamless iframes size correctly.")
window.onload = function () {
    window.iframe = document.getElementById("iframe");
    shouldBeTrue("iframe.seamless");

    // The nested iframe is also seamless.
    window.nestedFrame = iframe.contentDocument.getElementById("iframe");
    shouldBeTrue("nestedFrame.seamless");

    // Just like seamless-basic.html, the height from resources/square.html
    // should define the height of the nested seamless iframes.
    shouldBeEqualToString("window.getComputedStyle(iframe).width", "200px");
    shouldBeEqualToString("window.getComputedStyle(iframe).height", "100px");
}
</script>
